fix: fix loss performance
