+++ src/pages/main-page/ui/country-card.tsx
@@ -1,20 +1,52 @@
-import { memo } from 'react';
+import { memo, useCallback } from 'react';
 import { Country } from '@pages/main-page/api/types.ts';
+import { useLocalStorage } from '@pages/main-page/hooks/use-local-storage.ts';
 import classNames from 'classnames';
 
 import styles from './country-card.module.scss';
 
 interface Props {
 	country: Country;
-	isVisited: boolean;
-	onClick: (countryName: string) => void;
 }
 
-const CountryCard = ({ country, isVisited, onClick }: Props) => {
+const CountryCard = ({ country }: Props) => {
+	const [visitedCountries, setVisitedCountries] = useLocalStorage(
+		'visitedCountries',
+		'[]'
+	);
+
+	const visitedCountriesArray: string[] = JSON.parse(visitedCountries);
+
+	const isVisited = useCallback(
+		(countryName: string) => {
+			if (!visitedCountriesArray.length) return false;
+			return visitedCountriesArray.includes(countryName);
+		},
+		[visitedCountriesArray]
+	);
+	const handleCountryClick = useCallback(
+		(countryName: string) => {
+			if (isVisited(countryName)) {
+				setVisitedCountries(
+					JSON.stringify(
+						visitedCountriesArray.filter((country) => country !== countryName)
+					)
+				);
+			} else {
+				visitedCountriesArray.push(countryName);
+				setVisitedCountries(JSON.stringify(visitedCountriesArray));
+			}
+		},
+		[isVisited, setVisitedCountries, visitedCountriesArray]
+	);
+
 	return (
 		<div
-			className={classNames(styles.country, isVisited ? styles.visited : '')}
-			onClick={() => onClick(country.name.common)}
+			className={classNames(
+				styles.country,
+				isVisited(country.name.common) ? styles.visited : ''
+			)}
+			onClick={() => handleCountryClick(country.name.common)}
 		>
 			<img
 				className={styles.countryFlag}
